requirements.txt
git+git://github.com/nicolaiarocci/eve.git@develop
redis==2.4.13
Eve-docs==0.1.4
gevent==1.0.2
PyJWT==0.4.1
cffi==0.8.6
cryptography==0.7.2
enum34==1.0.4
pyasn1==0.1.7
pycparser==2.10
requests==2.7.0
rsa==3.1.4
six==1.9.0
wsgiref==0.1.2
